Adjust PR changes.
* Make parser take `&[u8]` not the AsRef<Vec<u8>>
* Move test from example to tests
* Simplify expressions a bit
* Add new version to Cargo.toml and CHANGELOG.md
