+++ image-blp/CHANGELOG.md
@@ -1,3 +1,6 @@
+# 1.2.0
+* Add `load_blp_from_buf`
+
 # 1.1.0
 * Fix parsing of malformed Dxtn images. Fix #2.
 * Update spin version
